+++ lib/ruby-lokalise-api/rest/queued_processes.rb
@@ -4,7 +4,7 @@ module Lokalise
   class Client
     # Returns all queued processes for the given project
     #
-    # @see https://app.lokalise.com/api2docs/curl/#transition-get-processes-get
+    # @see https://app.lokalise.com/api2docs/curl/#transition-list-all-processes-get
     # @return [Lokalise::Collection::QueuedProcess<Lokalise::Resources::QueuedProcess>]
     # @param project_id [String]
     # @param params [Hash]
@@ -14,14 +14,13 @@ module Lokalise
 
     # Returns a queued process for the given project
     #
-    # @see https://app.lokalise.com/api2docs/curl/#transition-get-process-get
+    # @see https://app.lokalise.com/api2docs/curl/#transition-retrieve-a-process-get
     # @return [Lokalise::Resources::QueuedProcess]
     # @param project_id [String]
     # @param process_id [String]
-    # @param process_type [String]
-    def queued_process(project_id, process_id, process_type = nil)
+    def queued_process(project_id, process_id)
       c_r Lokalise::Resources::QueuedProcess, :find,
-          [project_id, process_id, process_type]
+          [project_id, process_id]
     end
   end
 end