+++ spec/lib/ruby-lokalise-api/rest/queued_processes_spec.rb
@@ -2,8 +2,8 @@
 
 RSpec.describe Lokalise::Client do
   let(:project_id) { '803826145ba90b42d5d860.46800099' }
-  let(:process_id) { '4191a4e54cb6433193e75eed55359c4021e6aa91' }
-  let(:queued_process_id) { 'fb59937f47be1999b3643a7022d17bd46d807fad' }
+  let(:process_id) { '3b943469e6b3e324b5bdad639b122a623e6e7a1a' }
+  let(:queued_process_id) { 'ff1876382b7ba81f2bb465da8f030196ec401fa6' }
 
   describe '#queued_processes' do
     it 'returns all queued processes' do
@@ -13,7 +13,8 @@ RSpec.describe Lokalise::Client do
 
       expect(processes.branch).to eq('master')
       processes = processes.collection
-      expect(processes.count).to eq(2)
+      puts processes.count
+      expect(processes.count).to eq(8)
       expect(processes.first.process_id).to eq(process_id)
     end
   end
@@ -28,32 +29,14 @@ RSpec.describe Lokalise::Client do
     expect(process.type).to eq('file-import')
     expect(process.status).to eq('finished')
     expect(process.message).to eq('')
-    expect(process.created_by).to eq('20181')
+    expect(process.created_by).to eq(20_181)
     expect(process.created_by_email).to eq('bodrovis@protonmail.com')
-    expect(process.created_at).to eq('2020-05-11 11:17:52 (Etc/UTC)')
-    expect(process.created_at_timestamp).to eq(1_589_195_872)
-    expect(process.url).to eq('/api2/projects/803826145ba90b42d5d860.46800099/processes/file-import/4191a4e54cb6433193e75eed55359c4021e6aa91')
-  end
-
-  specify '#queued_process with process type' do
-    process = VCR.use_cassette('queued_process_detailed2') do
-      test_client.queued_process project_id, queued_process_id, 'file-import'
-    end
-
-    expect(process.process_id).to eq(queued_process_id)
-    expect(process.type).to eq('file-import')
-    expect(process.status).to eq('finished')
-    expect(process.message).to eq('')
-    expect(process.created_by).to eq('20181')
-    expect(process.created_by_email).to eq('bodrovis@protonmail.com')
-    expect(process.created_at).to eq('2020-05-14 10:18:03 (Etc/UTC)')
-    expect(process.created_at_timestamp).to eq(1_589_451_483)
-    expect(process.url).to eq('/api2/projects/803826145ba90b42d5d860.46800099/processes/file-import/fb59937f47be1999b3643a7022d17bd46d807fad')
-    expect(process.files.count).to eq(1)
-    file = process.files.first
-    expect(file['status']).to eq('finished')
-    expect(file['name_original']).to eq('rspec_async.yml')
-    expect(file['word_count_total']).to eq(1)
+    expect(process.created_at).to eq('2020-05-13 11:24:37 (Etc/UTC)')
+    expect(process.created_at_timestamp).to eq(1_589_369_077)
+    files = process.details['files']
+    expect(files.count).to eq(1)
+    expect(files[0]['status']).to eq('finished')
+    expect(files[0]['name_original']).to eq('test_async.json')
   end
 
   describe 'process status check' do
@@ -70,18 +53,6 @@ RSpec.describe Lokalise::Client do
 
       expect(reloaded_process.process_id).to eq(queued_process.process_id)
       expect(reloaded_process.status).to eq('finished')
-    end
-
-    specify '#reload_data with detailed' do
-      process = VCR.use_cassette('queued_process_detailed3') do
-        test_client.queued_process project_id, queued_process_id, 'file-import'
-      end
-
-      reloaded_process = VCR.use_cassette('queued_process_detailed3') do
-        process.reload_data
-      end
-
-      expect(reloaded_process.process_id).to eq(process.process_id)
     end
   end
 end